fix component module mispredict

+++ Content.Shared/Modules/Modules/AddComponentContainerModuleSystem.cs
@@ -1,10 +1,13 @@
 using Content.Shared.Modules.Components.Modules;
 using Content.Shared.Modules.Events;
+using Robust.Shared.Timing;
 
 namespace Content.Shared.Modules.Modules;
 
 public sealed partial class AddComponentContainerModuleSystem : EntitySystem
 {
+    [Dependency] private readonly IGameTiming _timing = default!;
+
     public override void Initialize()
     {
         base.Initialize();
@@ -15,11 +18,17 @@ public sealed partial class AddComponentContainerModuleSystem : EntitySystem
 
     private void OnAdded(Entity<AddComponentContainerModuleComponent> ent, ref ModuleAddedContainerEvent args)
     {
+        if (!_timing.IsFirstTimePredicted)
+            return;
+
         EntityManager.AddComponents(args.Container, ent.Comp.Components);
     }
 
     private void OnRemoved(Entity<AddComponentContainerModuleComponent> ent, ref ModuleRemovedContainerEvent args)
     {
+        if (!_timing.IsFirstTimePredicted)
+            return;
+
         EntityManager.RemoveComponents(args.Container, ent.Comp.Components);
     }
 }
